requirements.txt
futu-api
matplotlib
mplcursors
numpy
pytz
